Fix some issues regarding to security and logic

+++ src/main/java/com/example/devity/service/BookService.java
@@ -2,28 +2,27 @@ package com.example.devity.service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
 import org.springframework.stereotype.Service;
 
+import com.example.devity.configuration.AuthenticationContext;
 import com.example.devity.model.Book;
+import com.example.devity.model.User;
 import com.example.devity.repository.BookRepository;
+import com.example.devity.repository.UserRepository;
 
 @Service
 public class BookService {
 
   private BookRepository bookRepository;
+  private UserRepository userRepository;
 
-  public BookService(BookRepository bookRepository) {
+  public BookService(BookRepository bookRepository, UserRepository userRepository) {
     this.bookRepository = bookRepository;
-  }
-
-  public List<Book> findAll() {
-    Iterable<Book> bookIterable = this.bookRepository.findAll();
-
-    return StreamSupport.stream(bookIterable.spliterator(), false)
-        .collect(Collectors.toUnmodifiableList());
+    this.userRepository = userRepository;
   }
 
   public Book findById(long id) {
@@ -31,7 +30,7 @@ public class BookService {
         .orElseThrow(() -> new IllegalStateException((String.format("book with this id = %d not found", id))));
   }
 
-  public List<Book> findByDetail(String description, String title) {
+  public List<Book> findAll(String description, String title) {
     List<Book> books = new ArrayList<>();
     if (description == null && title == null) {
       Iterable<Book> bookIterable = this.bookRepository.findAll();
@@ -50,10 +49,18 @@ public class BookService {
   }
 
   public Book create(Book book) {
+    String username = AuthenticationContext.getAuthContext().getName();
+    User systemUser = userRepository.findByUsername(username).orElseThrow();
+    book.setUser(systemUser);
     return bookRepository.save(book);
   }
 
   public void delete(long id) {
-    bookRepository.deleteById(id);
+    Book book = findById(id);
+    if (!Objects.equals(AuthenticationContext.getAuthContext().getName(), book.getUser().getName())) {
+      bookRepository.deleteById(id);
+    } else {
+      throw new IllegalStateException("user can delete only there own books");
+    }
   }
 }
